campus-ai-backend/pyproject.toml


[project]
name = "hackathon-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Umar Hayat", email = "umarhayat.coder@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-google-genai (>=2.1.12,<3.0.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pymongo (>=4.12,<4.15)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pypdf (>=6.0.0,<7.0.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "motor (>=3.7.1,<4.0.0)",
    "langgraph-checkpoint-mongodb (>=0.2.0,<0.3.0)",
    "openai (>=1.108.1,<2.0.0)"

    
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
